Enrich an overpass query for column output

Enrich an overpass query for column output

enrich_opq(
  dataset,
  name = NULL,
  key = NULL,
  value = NULL,
  type = "points",
  measure = "spherical",
  r = NULL,
  kernel = "uniform",
  reduce_fun = sum,
  control = list(),
  .verbose = TRUE,
  ...
)

add_bbox(dataset, r, control)

add_feature(opq, key, value)

add_type(opq, type)

add_measure(opq, measure)

add_kernel(opq, kernel, r, reduce_fun, ...)

Arguments

dataset
target sf dataset to enrich with this package

name
name of the enriched column

key
target OSM feature key to add, see osmdata::add_osm_feature()

value
target value for OSM feature key to add, see osmdata::add_osm_feature()

type
character the osm feature type or types to consider (e.g., points, polygons), see details

measure
character the measure metric used, see details

r
The search radius used by the kernel function.

kernel
function the kernel function used, see details

reduce_fun
The aggregation function used by the kernel function to aggregate the retrieved data objects

control
The list with configuration variables for the OSRM server. It contains timeout, defining the number of seconds before the request to OSRM times out, and memsize, defining the maximum size of the query to OSRM.

.verbose
bool whether to print info during enrichment

...
Additional parameters to be passed into the OSM query, such as a user-defined kernel.

opq
overpass query that is being enriched

See also