Enrich an overpass query for column output

Enrich an overpass query for column output

enrich_opq(
  dataset,
  name = NULL,
  key = NULL,
  value = NULL,
  type = "points",
  distance = "spherical",
  kernel = "uniform",
  .verbose = TRUE,
  ...
)

add_bbox(dataset, ...)

add_feature(opq, key, value)

add_type(opq, type)

add_distance(opq, distance)

add_kernel(opq, kernel, ...)

Arguments

dataset
target sf dataset to enrich with this package

name
name of the enriched column

key
target OSM feature key to add, see add_osm_feature

value
target value for OSM feature key to add, see add_osm_feature

type
character the osm feature type or types to consider (e.g., points, polygons), see details

distance
character the distance metric used, see details

kernel
function the kernel function used, see details

.verbose
bool whether to print info during enrichment

...
arguments passed to the kernel function

opq
overpass query that is being enriched

Details

Type represents the feature type to be considered. Usually this would be points, but polygons and multipolygons are also possible. This argument can also be a vector of multiple types. Non-point types will be converted to points using the st_centroid function from the sf package (NB this does not necessarily work well for all features!) Available options are:

points

lines

polygons

multilines

multipolygons

Distance represents the metric used to compute the distances between the rows in the dataset and the osm features. Duration represents the metric that indicates the average duration to cover the distances between the rows in the dataset and the osm features. The following metrics are available in this package, assuming that the OSRM server is setup as suggested in our guide at: https://github.com/sodascience/osmenrich_docker:

spherical ("as the crow flies")

distance_by_foot

duration_by_foot

distance_by_car

duration_by_car

distance_by_bike

duration_by_bike

Kernel is a kernel function from the osmenrich package to be used in weighing the features and the radius/distance where features are considered. For simply counting the number of occurrences within a radius, use kernel_uniform with radius r.

See also